feature/ header의 masthead 영역을 컴포넌트로 분리
