+++ src/assets/scss/_layout.scss
@@ -1,0 +1,3 @@
+$border-radius: 4px;
+$HEADER-HEIGHT: 64px;
+$screen-main-max-width: 900px;
