+++ src/assets/scss/_theme.scss
@@ -1,8 +1,5 @@
 $bg-color: #eeeeee;
 $SURFACE-COLOR: #ffffff;
-$border-radius: 4px;
-$header-height: 60px;
-$screen-main-max-width: 900px;
 
 @import url('https://fonts.googleapis.com/css2?family=Nanum+Gothic&family=Roboto&dispay=swap');
